rules: add custom per-group mapping for level3(ralt_alt)
ralt_alt is a special mapping: "the right Alt key never chooses the
third level." To achieve this, the key is defined for all four possible
groups.

This causes an issue with the libxkbcommon parser which extends the
keymap to be 4 groups wide thanks to this key, see
https://github.com/xkbcommon/libxkbcommon/issues/262
That is arguably a bug in libxkbcommon whose behavior differs from
xkbcomp - xkbcomp clamps to the number of layouts.

But defining the key this way is wrong, no other keys is defined for
multiple groups on that purpose. The correct way to achieve this
functionality is to define the symbols for group1 only and then adjust
the rules so they automatically apply to all groups currently present.