symbols/np: add parentheses and period to the basic Nepali layout
The "॰" that was on AE10 is never used in Nepali, so the second level
of AE09 and AE10 can be used for parentheses instead, as is custom.

Instead of "॰", the Nepalese use the period (.), so this character is
put on the second level of AB09, pushing the "॥" that was there to the
second level of AC09, which is possible because the "ळ" that was there
is not used in Nepali.

Fixes #251.

Reported-by: Pranav Bhattarai