revert !238, to get 'less' and 'greater' back on many layouts
Five months ago, '<' and '>' were replaced with '‹' and '›' on many
layouts by commit b17b586087.  It seemed like a logical thing to do:
put single guillemots on the fourth level when the double ones are
on the third.  However, this made '<' and '>' untypable for a bunch
of people because most keyboards do not have a LSGT key and their
layouts did not contain any other key with 'less' or 'greater'.

There were reports for Latin American, Turkish, Spanish, Italian.
And Norwegian and Portuguese are also affected, from a quick check.

So, revert the change, also because these single guillemots are
useful to hardly anyone.

Fixes #286.

Reported-by: Eduard Tolosa