symbols/is: add bar on third level of P, for keyboards without LSGT key
The third and fourth level of key AD10 contained eth and ETH, but on
Icelandic layouts these letters have their own dedicated key (AB10).
So, replace the duplicate eth with bar, and ETH with Greek pi (π).

The Greek pi was chosen for the fourth level because it is related to
"p", it reinforces the mnemonic for "pipe" (bar), and there is another
Greek letter (Ω) at the other end (AD01).

Fixes #289.

Requested-by: David Olafsson