symbols/is: remove an obsolete comment and some unneeded comments
The comment about "corrected" is out-of-date because most of the changes
that were made by Bjarni have been undone.  And the extensive historical
text is present in macintosh_vndr/is; no need to duplicate it here.

Also, move the TLDE key to its normal position, and drop the definition
of the SPCE key as it doesn't change anything over the default.