replace the non-working 'horizconnector' with a bullet (U+2022)
Key symbol 'horizconnector' has not been working for more than eleven
years.  In all that time no one has complained, so... it seems no one
was using that character.

So, replace it with something that works, and that is more mnemonic
(having a comma or a period on one of the first two levels, and the
centered period to its right): the bullet character, U+2022.

Fixes #301.

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>