symbols/eu: restore 'section' to the fourth level of the S key
The uppercase Ssharp (ẞ) is nearly useless, and *when* needed, it is
obvious how to compose it: <Compose> <S> <S>.  The section symbol (§)
looks more useful, and it is not obvious how to compose it -- I would
try <Compose> <s> <s>, because it looks like two esses intertwined,
but that of course doesn't work.

Together with commit 2d9cddfc64 from November last year, this completes
the full revert of commit 08ef288b78 from October last year.

Fixes #271.