make the Win+Space toggle work in every group by adding a rule for each
In order for an option to override the multiple levels of a key that
any of the layouts defined, the option must be applied to each group
separately AND the key type must be specified explicitly for group 1.

(Also improve a comment and adjust the indentation to match others.)

Fixes #90.

Reported-by: Oliver Majersky