symbols/ara: put the vertical bar back on Shift+Backslash
Commit 83982bd6a0 from four years ago replaced the vertical bar on the
second level of the BKSL key with an ellipsis, leaving the vertical bar
accessible only via the LSGT key.  However, most Arabic keyboards don't
have an LSGT key, meaning that the vertical bar became untypable.  Fix
this by restoring the vertical bar to the second level of the BKSL key,
and moving the ellipsis (when possible) to the second level of the LSGT
key, displacing the almost useless brokenbar symbol.