rules: make the RCtrl+RShift toggle work by adding a per-group rule
For some reason, when there is only a general rule for an option,
it gets applied (always? often?) only to the first group, which is
problematic especially for group toggles, because they are meant
to work in every group.  Avoid the problem by explicitly defining
a rule for the RCtrl+RShift toggle for every group.

This fixes issue #252.

Originally-reported-by: Leonard Lausen