rules: make the Rctrl and Caps switches for the fifth level selectable
The Rctrl switch for the fifth level has been available for seventeen
years (since commit 075eed027d), and the Caps switch for five years
(since commit bb80698c2b), but were not exposed in the rules files.

Also, restore the Menu switch for the fifth level that existed for a
while a year ago, and make it selectable.

This fixes issue #314.

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>