rules: correct three descriptions to say "second" instead of "last" layout
Contrary to what the symbol `ISO_Last_Group` seems to imply, it does
not select the last group that the user defined but the second group.
Adjust the descriptions to reflect that.

Also make a fourth description more accurate: for lctrl_lwin_rctrl_menu
any Ctrl key will do -- which will be relevant when using CapsLock as
Ctrl, for example, or the right Ctrl key as fifth-level chooser.

Also adjust the comments in the symbols file accordingly.

This addresses issue #315.

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>