symbols/tr: delete two definitions of BKSL that are overridden by AC12
(AC12 is the same key as BKSL, just an alias of it.)

Also sort a different AC12 in its right place, in symbols/pl.

And delete the use of an invalid key name, AD13, in symbols/my.
(The intended key, BKSL, stays defined anyway, as it is defined
by default in symbols/pc.)