symbols/ru: make Shift+6 produce ^ (not ‸) in the phonetic Mac Layout
The intent of commit db988e4b75 seems to have been to make Shift+3 to
Shift+6 produce the usual series of # $ % ^ symbols, but it mistakenly
used the human name "caret" for ^, instead of the "asciicircum" name
as defined in keysymdef.h.  The keysymdef.h file defines "caret" as ‸
(U+2038), a distinctly different symbol.

This amends issue #191.

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>