add a "Both Alts; AltGr for level 3" option for switching between groups
This allows the user to switch between layouts with the two Alt keys
(to either the previous or the next group) *and* to use the right Alt
key as third-level chooser.  In the bargain it avoids either Alt key
from producing a different symbol upon key release than upon press.

This improves the fix for issue #316.

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>