symbols/mm: remove unneeded upper-level definitions that copy lower ones
(There is no need to specify the symbols for levels 3 and 4 when
they are the same as those for levels 1 and 2.)

Also, put the BSKL key in its usual order, and the inclusion of the
ralt_switch at the end of a layout (where it usually is), and the
inclusion of the tilde_switch at the beginning of a layout, where
the TLDE key normally is.  And rename the fragment to tilde_latch,
because it behaves differently than the ralt_switch.