symbols/in: drop four definitions of Right Alt key that get overridden
These four definitions assign two symbols to the key and immediately
afterwards include the 'ralt_switch' from 'level3', which forces the
key to a single level and sets ISO_Level3_Shift as the only symbol.

(Also, the four definitions assigned Mode_switch to the key.  But this
only makes sense when the user defines at least two groups, which is
something that a layout cannot know.)