symbols/bt: update the Dzongkha layout to the revised standard of 2009
The original Bhutanese standard keyboard layout from 2000 was revised
in 2009 -- https://en.wikipedia.org/wiki/Dzongkha_keyboard_layout.

The changes are mostly in the fourth level, replacing some ASCII symbols
with code points that had been assigned in the meantime.  It also swaps
colon and semicolon to be in their usual US-keyboard arrangement, and
replaces square brackets with parentheses -- although, the standard is
unclear there: it shows parenthesis glyphs but gives U+0F28 and U+0F29
as code points, which are the Tibetan digits 8 and 9, which doesn't make
sense; so, most probably code points U+0028 and U+0029 were meant here,
which are the normal parentheses.

The unassigned code point U+0F48 is kept on the second level of AE04,
the extra parentheses are kept on the fourth level of AE09 and AE10,
and the single and double guillemots are kept on the third and fourth
levels of AC05 and AC06, even though the standard reserves these spots
and has them empty.  It is unclear whether the extra characters were
part of the earlier standard or fanciful additions by the submitters
of the layout in 2004.

The fourth level of AD05 should produce U+0F8B plus U+0F99, but XKB
cannot do that -- a compose sequence will be needed to achieve this,
some day.  Until then, people can achieve the combination by typing
Shift+<AD12> after Shift+AltGr+<AD05>.

The definition of the BKSL key is moved to below the AE keys,
for easier comparison with the diagrams of the standard.

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>