symbols/de: replace the words "Mod3" and "Mod4" in comments of Neo layout
In XKB, the keywords Mod1 to Mod5 have a specific meaning: they are
the identifiers of actual modifier bits in the keyboard state.  It
was therefore confusing to see the comments for the Neo layout use
the words "Mod3" and "Mod4" with a different meaning: as abstracted
names for certain keys.  When grepping the XKB files for Mod[1-5],
one wants to find only occurrences that refer to the actual modifier
bits, and not occurrences that mean something else.  So, reword the
latter occurrences to avoid false positives.

[To document the confusion: the Neo "Mod three" key affects the Mod5
modifier bit (LevelThree), while the "Mod four" key affects the Mod3
modifier bit (LevelFive) when just setting it, and Mod2 (NumLock) when
locking it.]

Additionally make a few esthetic tweaks.

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>