rules/compat: retire a long-obsolete code for Serbia + Montenegro
The Union of Crna Gora and Srbija ceased to exist sixteen years ago
and its country code (CS) was deleted from ISO_3166 the next year.
(See, for example, https://en.wikipedia.org/wiki/ISO_3166-2:CS.)

Nowadays, when a user does 'setxkbmap cs', they most likely want to
set a Czech layout, not a Serbian layout, so it will be much better
to fail with an error message in that case than to set a Cyrillic
layout that is both unexpected and makes it impossible for the user
to issue a command to get back a usable keyboard.