map 'ISO_Level5_Shift' by default to the Mod3 modifier through <MDSW>
Assign the ISO_Level5_Shift symbol to the <MDSW> key and map this key
to the Mod3 modifier, so that, when the `action` part of an `interpret`
rule for ISO_Level5_Shift wants to set a virtual modifier (LevelFive,
normally), it will actually set Mod3.

This new default arrangement makes the fifth-level choosers (the <RALT>
and <LSGT> keys) work for the Neo layout (and its derivatives) also when
this layout is not the first group.

The <MDSW> key can be repurposed because "Mode switching" works fine
without having the symbol 'Mode_switch' attached to some modifier:
the `action` part of the relevant `interpret` rule simply increases
the group number with 1, and that is enough, no need to also set Mod5.

This fixes issue #322.

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>