Split Indonesia (Arab Melayu) and Indonesia (Arab Pegon)
"Arab Melayu" layout was renamed to "Arab Pegon". However, Arab Melayu and
Arab Pegon writing systems are actually slighly different, so that I think
this was a bug. The fix is to revert Arab Pegon name to Arab Melayu and
introduce a new layout named Arab Pegon which is copied and modified from
Arab Melayu. For this purpose, the Arab Melayu 'phonetic' and 'phoneticx' are
renamed to 'melayu-phonetic' and 'melayu-phoneticx' respectively, while the
Arab Pegon one is named 'pegon-phonetic'.

Additionally, both in Arab Melayu and Arab Pegon, I also add "theh" letter to
the position of AltGr+Shift+S, because phonetically "theh" is closer to S,
its current position at V is just a shortcut, imitating the position of "theh"
in Arabic phonetic layout.