symbols/ua: correct the type of a key, to being four-level alphabetic
The key type needs to be specified explicitly because the server does
not recognize "idotless" and "I" as a lowercase-uppercase pair and thus
does not assign the key type automatically.  And anyway, it is necessary
to override the type assigned by the included Turkish layout.

(The key specified the wrong key type probably because the definition
was simply copied from the relevant Turkish layout and only the upper
symbols were changed.)

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>