rules/compat: delete an alias that should no longer be needed
Eighteen years ago, in commit 1bdb65affc, the inaccurately named
option 'shift_toggle' was renamed to the better 'shifts_toggle',
and 'shift_toggle' was added as a convenience alias for existing
users of the option.  For how long should such an alias be kept?
Ten years seems more than enough, so it's time to drop this one.

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>