symbols/no: add non-dead grave and such to Sami layout, and reshuffle
When working at a computer, it's nice to have a non-dead grave (`) and
tilde (~) and caret (^).  They are given the places they have in the US
layout, just two levels higher, like less-than (<) and greater-than (>).

The division symbol (÷), plusminus (±), and notsign (¬) are put in the
same spot as in the basic Norwegian layout.  The spots for multiply (×),
less-or-equal (≤) and greater-or-equal (≥) logically follow.

The guillemots are moved down to the third level, to be more accessible,
as they are the main quotation symbols in Norway.  The other quotation
marks are logically distributed around these, in the same manner as the
curly braces are “wrapped around” the square brackets on the upper row.

The single guillemots (‹ and ›) on the fourth level are rather useless
symbols, but they reinforce the logic of the arrangement, so that it
becomes easier to remember.  Furthermore, ‘vasen’ in Finnish, ‘vänster’
in Swedish, and ‘venstre’ in Norwegian mean "left", which might help to
remember that the left guillemots are on the 'V' key.

References: https://en.wikipedia.org/wiki/Quotation_mark and
https://jakubmarian.com/map-of-quotation-marks-in-european-languages/.

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>