symbols/ee: add extra Ä Ö Ü Õ to US variant, to ease two-handed entry
The letters with diaeresis (Ä, Ö, Ü) are added on the free upper levels
at the base position of the fingers of the left hand, and the latter two
also on the upper levels of their root letter.  As that leaves no room
for Õ, it gets pushed to the neighboring key (P), and gets assigned to
Q too (for mnemonics: both are a circle with a squiggle), and also to X
(the closest key, and because Š is related to Ž, and Ö to Õ).

This partially addresses issue #106.

Requested-by: Ander Punnar

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>