symbols/us: uncomment the nine keys in the international Dvorak layout
When a layout redefines almost all keys inherited from an included
layout, it is both faster and clearer to drop the include and just
define all keys in the layout itself.

Also, <AD12> and <AC11> were wrongly commented out, thus depriving
the Dvorak-international users of access to a few useful symbols.

This fixes issue #358.

Reported-by: J. McWilliams

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>