symbols/us: remove a pointless include from the basic Workman layout
All alphanumeric keys in this basic Workman layout are defined with
just two levels, so it makes no sense to include a third-level chooser.
And the one key that had four levels (CAPS) had the same symbols on the
upper levels as on the first, so those levels provided nothing extra,
and are now removed.

This closes issue #362.

Reported-by: Jing Wang

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>