Remove the obsolete RSTU keyboard layout, fix the name of RUU
* The RSTU 2019-91 standard was developed in the Ukrainian SSR back in 1991
in consideration of electronic typewriters and therefore was not officially
used in computer technology. After the collapse of the Soviet Union,
in independent Ukraine in 1996, the DSTU 3470-96 standard has been adopted.
Despite the fact that the RSTU layout was added to this repository in 2004,
officially this standard has been dead since 1991. Also, it is almost
impossible to connect old keyboards with 5-pin DIN connectors to modern
computers.

* The RUU layout is slightly renamed, as the official name of the country
is Belarus, not Belorussia.