symbols: replace some mistaken keypad keysyms with the intended symbols
The KP_n symbols on the second level in the 'azerty' layout for Cameroon
still make some sense: the keys get the type FOUR_LEVEL_KEYPAD so that
the digits will be produced when NumLock is engaged.  But in the 'dvorak'
layout the KP_n symbols were on the first level, which would lead to the
symbols from the second level to be produced when NumLock is engaged --
which is not what NumLock is supposed to do.  (Truly, NumLock is meant
only for the numeric keypad, but... let's not disturb current users.)