Add Persian digits to the Windows Persian variant
In Windows, the Persian layout has Latin (ASCII) digits only, and
some programs convert them to Persian digits as they are typed.
This type of conversion does not exist on most systems, including
freedesktop-based systems, making it impossible to type Persian digits
when using the Windows variant of the Persian keyboard layout.
This commit adds Persian digits to the third level of the number keys
on row E (AE01 - AE10).