symbols/de: correct the key type of the letter keys in the E1 layout
In the E1 layout, the keys with letters on the first two levels
contain symbols (not letters) on the third level, and nothing on
the fourth level.  So, CapsLock should not affect the third level,
otherwise AltGr plus a letter key would do nothing when CapsLock
is engaged.

This fixes issue #333.

Problem existed since the E1 layout was introduced three years ago
in commit b9063591ec.

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>