symbols/gr: harmonize and condense whitespace, unwrap some definitions
Also, remove the marker `alternate_group` -- a leftover from the time
that the upper levels were meant to be accessed in a subsequent group
via `Mode_switch`.

Furthermore, replace `NoSymbol` with `any`, and amend the graphical
depiction of the Extended layout (which, by the way, is less extended
than the basic layout).