force the CAPS key to have one level (for options that affect that key)
In olden times, there were no layouts that redefined the CAPS key,
so having options that overwrote the single symbol that it contained
worked fine.  However, some modern layouts, like Colemak and Workman,
define the CAPS key to have multiple levels, which causes some of the
XKB options to fail to overwrite the upper levels of that key.

Indirectly-reported-by: Vinícius dos Santos Oliveira

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>