elide an unexposed non-breaking-space option that was used just once
The option 'nb2zwnj3s' was added twelve years ago in commit 0a247a4cdd,
for the sole benefit of the Sri Lankan phonetic layout.  But using an
include does not make things clearer, especially not when its name is
hard to grasp.  Also, the option was never exposed in any rules file,
and was thus not really accessible for anyone.  So: just drop it.