rules: make the `Shifts` group toggle override any layout's Shift keys
In order to make the option `grp:shifts_toggle` effective also with
the few layouts that define the Shift keys too, the option should be
applied to each group separately, and the option should specify the
key type explicitly.

This fixes issue #380.

Problem existed since the T3 layout was added in 2013,
and worsened when the E1 layout was added in 2020.

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>