Add U+06C0 to the Persian layout
In Persian, an isolated form of the letter Ye is added after nouns
that are connected to adjectives, and if the noun ends with the
letter He, it is often merged with the following Ye, into the letter
He with Ye above (U+06C0).
This character exists on the second level of AC05 in the Persian
layout of Microsoft Windows, but not in the standard layout.
This commit extends the standard layout to include U+06C0 on the
third level of AC05.