symbols/et: use dead-vowel symbols instead of mistaken Unicode codepoints
When twelve years ago, in commit c28367a15, the code points 0x100FE67
and further were added for "dead a" and such, the intention was that
they were raw key codes (0xfe67 and further), not Unicode codepoints.
See the original patch that requested adding very similar key codes
to xorgproto three years earlier:

   https://bugs.freedesktop.org/attachment.cgi?id=18679

The compose sequences in libX11 have been updated to use the dead-vowel
symbols instead of the erroneous Unicode codepoints (most of which are
valid codepoints for certain characters, that thus became unusable in
the Amharic locale), so now it's time to update the keyboard layout.

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>