symbols/pc: map the Super modifier to the LWIN/RWIN keys, not to symbols
When a modifier is mapped to a symbol, then it will effectively be
mapped to the first-found key that contains the relevant symbol,
and that key might not be the key that the modifier is usually
mapped to, leaving this latter key without a mapping.

So, map the Super modifier to `<LWIN>` and `<RWIN>` instead of to
`Super_L` and `Super_R` -- the Windows keys will then not lack their
Super modifier when some other key is made into a Super key too.

This fixes issue #81.

Reported-by: Josh Triplett

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>