symbols: map three real modifiers to keys instead of to symbols
Mapping a modifier to a symbol had the tiny advantage of not needing
to explicitly specify a modifier mapping when putting the symbol onto
some other key through some option.  But it had the disadvantage of
risking to lose the modifier mapping for the original key when the
other key was just an addition and not a replacement.

This generalizes the fix from the previous commit.

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>