revert commit 1cb7181936 plus the offending part of a1d41a4ef6
Mapping modifiers to keys instead of to symbols causes problems when
those keys carry symbols that are associated with *other* modifiers,
as this leads to two or more modifiers getting linked together and
thus activated in unison -- which one normally doesn't want.

This fixes issue #382.

(Issue #81 will have to be addressed in some other way.)

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>