improve the names of two Czech and two Slovak variants
Nineteen years ago, commit 8e1e2d8195 baptized these four variants with
the words "extended backslash".  Nine months later, commit 999e981f4c
synced these names to the relevant symbol files.

But just two days later, commit 3ad7cee7c5 changed one of the names to
read "with <\|> key", in response to https://bugs.freedesktop.org/2319.
(A week later, commit 1b8556fe6f synced this name to the XML file.)
This gave one of the four very similar variants a deviant name.

The string "with <\|> key" is difficult for translators.  For a long
time I have thought that "<" and ">" were the relevant symbols, and
that the "|" was an OR symbol, escaped by the "\".  In other words:
I thought that the string referred to the LSGT key.  :/

What the four variants actually do is: put the backslash and vertical
bar characters on the BKSL key (replacing the dead diaeresis and the
apostrophe), for those keyboards that lack the LSGT key, to make typing
"\" and "|" more straightforward -- these symbols are present also on
the third level of the Q and W keys.  So what those variants do is:
add an *extra* backslash and vertical bar character to the layout,
for easier access, at the cost of having the apostrophe accessible
only via AltGr and the dead diaeresis only via Shift+AltGr.

So... name all four variants more fittingly and in the same manner.

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>