add a layout for the Gothic script
This Gothic layout is put in a new symbols file, called 'ancient',
so that there will be a place for other layouts for antique scripts.

The name "ancient" is chosen because that is how Gothic is marked
in https://en.wikipedia.org/wiki/List_of_ISO_639-2_codes.

This fulfills issue #55.

Proposed-by: Tim Nelson
Original-patch-by: Tim Nelson

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>