symbols: replace the remaining dead symbols in most "nodeadkeys" variants
This leaves one dead symbol in the Finnish "nodeadkeys" variant,
and five dead symbols in the Hungarian one, but... they are all
on the fourth level, so... let them be.

The upper-level dead symbols are replaced with the marginally useful
ezh (`ʒ`), dagger (`†`), radical (`√`), and ellipsis (`…`).

This partially addresses and closes issue #339.

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>