symbols/se: remove a variant that does not differ from its parent
The "rus_nodeadkeys" variant included se(nodeadkeys), but then it
also included se(rus) which includes se(basic) which overwrites all
the non-dead symbols with dead symbols again.

In the eighteen years that these Russian variants existed no one has
complained about the dead keys in the "rus_nodeadkeys" variant, so...
just delete it, as it is identical to the "rus" variant.