symbols/us: Add `periodcentered`, `dead_stroke` to us-intl variants
`dead_stroke` is required to type four letters (eight characters) of
Latin Extended-A. Additionally, `periodcentered` is a standard
alternative multiplication symbol within the context of algebra in
the United States. Both of which are put on the 3rd and 4th levels of
`x` with the nomenclature "multiply" and "cross-out". However, for
variants whose `x` position is already occupied, namely us-altgr-intl
and capalx-intl, they are instead given the nearest unused key
position.