symbols/fr: replace a duplicated symbol with a missing one in Mac layout
The French Mac layout contained an uppercase Ÿ on the fourth level of
`TLDE` *and* on the fourth level of `AD06`, whereas the lowercase ÿ was
missing.  Even though Ÿ and ÿ are extremely rare in French, this makes
no sense, so... replace that first copy of Ÿ with ÿ, because when this
ÿ is not what the user wanted, then the most logical place to search
for Ÿ is the upper levels of `AD06`, the Y key.

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>