symbols/ru: add a duplicate symbol on upper level of key in Yakut layout
It seems that many speakers of Yakut will use the Cyrillic_softsign
in lieu of the Cyrillic_shha when the keyboard layout doesn't support
the Yakut language.  When the layout *does* support all Yakut letters,
these speakers are somewhat likely to try and type the Cyrillic_shha
via the upper levels of the Cyrillic_softsign key.

Requested by: Prokopii Ivanov

This supersedes merge request !507.

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>