symbols/se: replace duplicated upper-level symbols with semi-useful ones
This especially replaces the `§` symbol on the fourth level of `AD02`
because it already exists on the first level of `TLDE`, which makes it
somewhat likely to be used in a shortcut, which can lead to the other
symbols on `AD02` getting filtered out in some desktop environments.

(The latter is of course a problem in those desktop environments,
but there's is no harm in preventing the problem from occurring
by eliminating the duplicate symbol.)

It further replaces a duplicated `¢` with `‰`, a useless `⅝` with `√`
(because ⅛, ⅜ and ⅞ are absent), a duplicate `@` with `ω`, a useless
`ſ` with `σ`, the duplicated `§` with `Σ`, a duplicate `®` with `™`,
two duplicated thorns with `π` and `Π`, a duplicated `&` with `†`,
a duplicate `©` with `🄯`, and a duplicate dead abovedot with `…`.

This fixes issue #388.

Reported-by: Markus Hartung

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>