symbols/ara: add the Mac phonetic layout
This layout is an improvement over Buckwalter which has dead keys as well as
missing symbols on the second level, and requires the Shift key for common
letters such as alef, hah and ain. The Mac layout has no dead keys and makes
use of four levels, with a straightforward mapping where the first level
matches vocalization or letter shape, and the second provides the dotted form.
Higher levels provide diacritics and other symbols. Providing a layout already
available on a major platform allows users to move between platforms with ease.