man: use text entries for the option descriptions
These may exceed the terminal width by quite a bit, so let's add those
as text entries that can be wrapped. To have this work properly, we need
to expand ('x') this column into the available space - whatever roff
decides that space is. This way at least it doesnt try to wrap after
single word or two.

Output of a table entry changes from
  foo some description
to this:
  foo T{
  some description
  T}

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>