symbols/{al,pl}: replace another duplicated symbol with semi-useful one
Two years ago, commit 9aceb34058 replaced a duplicated `Ł` with a `§`
that was moved from the `S` key.  It did this not only for the Latin
type1 layout but also for the type3 layout.  But two of the three
layouts that include latin(type3) already force `§` to be on the
fourth level of `S` (which is obviously more mnemonic), so the `§`
on the fourth level of `W` is a pointless duplicate for those two
layouts.  Replace the duplicate `§` with the semi-useful `Σ`.

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>