man: fix the last lines of the various tables
For the Layouts, we use the XSL position()!=last() check to add the
divider lines for all but the last element - this gets rid of the empty
box after the custom layout. Also remove a superfluous empty lines at
the top of each indidviual country code box.

For all the other tables, remove the added line break before .TE which
removes the last empty line in each table.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>