symbols/cz: change an include, reduce a full definition to nine overrides
The recent Czech `winkeys` variant is just a slight variation of the
default Czech `basic` layout, differing only in the upper levels of
nine keys on the righthand side of the keyboard.

So... let the `winkeys` layout include the `basic` layout and then
just redefine the nine relevant keys, instead of doing a complete
definition of all keys.

Also, shrink a verbose comment and delete a repeated one.

Also, this `winkeys` layout only *mostly* conforms to what is printed
on actual keyboards in Czechia: the example image that was linked to
has `comma` and `period` on level 1 and level 3 of AB08 and AB09, and
`less` and `greater` on level 4, but this layout has levels 3 and 4 on
those keys swapped.  The latter is arguably somewhat more convenient,
so... let it be.  But the original commit (6d9f17da14) should have
mentioned this discrepancy.

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>