meson: run the pytest test suite as part of meson test
If pytest is available, run the test suite during meson test.
This requires a bit of fiddling around because the test suite expects
the XKB directory to be properly laid out but that doesn't happen with
meson until install. Luckily all we need to do here is copy our KcCGST
directories over to the build directory and then we can use that as
XKB_CONFIG_ROOT for pytest.

An optional hook for pytest-xdist is integrated too - on my machine the
test run goes from 25s to 10s with xdist on -n auto. This does require
a conftest.py so we don't hog all the cores available on the CI runners.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>