symbols/lt: add a `{` and a `»` to make the "ibm" layout more consistent
The `]` and `}` on level 3 of AE09 and AE10 (inherited from the included
"latin") were meant to match with `{` and `[` on AE07 and AE08, but the
"ibm" layout overwrites the latter two with `&` and `*`, leaving the `]`
and `}` unmatched.  Match this `}` by replacing the unpaired `]` on AE09
with `{`.  (A pair of `[` and `]` are present on level 3 of AD11/AD12.)

Furthermore, the "ibm" layout had only a `«`, not any `»`, whereas the
unpaired `<` that was on AB01 is already present on two other keys (LSGT
and AB08) where it *is* paired.  So, replace that lonely `<` with `»`.

Also, drop the "LST 1205-92" from the name of the layout, to avoid the
impression that it faithfully follows the standard, and put the "IBM"
part between backquotes, to indicate it's not really the IBM layout.

This addresses and closes issue #398.

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>