symbols/in: rename a layout, by mentioning the script instead of the font
"Eeyek" is the name of a font [1], not the name of the relevant script.
The script is called "Meitei" [2], like the language and the people.
The official name of the language is "Manipuri", and its code `mni`.

  [1] https://sources.debian.org/patches/fonts-eeyek/
  [2] https://en.wikipedia.org/wiki/Meitei_script

So, rename and re-identify the layout accordingly, and add an alias
from `eeyek` to `mni`.

Also, drop the third-level chooser, as the layout does not contain
anything above the two base levels.

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>