symbols/{se,us}: correct the comments of two phonetic Russian layouts
Eighteen years ago, in response to https://bugs.freedesktop.org/2582,
commit abf223b22c added two phonetic layouts for typing Russian on a
Swedish and on an American keyboard.  The original patch had all the
Swedish and American letters on the third and fourth level, but for
the most part only the first two levels were committed.  The comments,
though, were not adjusted to what was committed.  Do that now.

Also, remove a mistaken key type that made the upper-level symbols
react to the Lock modifier, which they shouldn't.

Signed-off-by: Benno Schulenberg <bensberg@telfort.nl>